xdg-shell: split into multiple files
