+++ types/meson.build
@@ -9,6 +9,11 @@ lib_wlr_types = static_library(
 		'seat/wlr_seat_pointer.c',
 		'seat/wlr_seat_touch.c',
 		'seat/wlr_seat.c',
+		'xdg_shell/wlr_xdg_popup.c',
+		'xdg_shell/wlr_xdg_positioner.c',
+		'xdg_shell/wlr_xdg_shell.c',
+		'xdg_shell/wlr_xdg_surface.c',
+		'xdg_shell/wlr_xdg_toplevel.c',
 		'xdg_shell_v6/wlr_xdg_popup_v6.c',
 		'xdg_shell_v6/wlr_xdg_positioner_v6.c',
 		'xdg_shell_v6/wlr_xdg_shell_v6.c',
@@ -42,7 +47,6 @@ lib_wlr_types = static_library(
 		'wlr_wl_shell.c',
 		'wlr_xcursor_manager.c',
 		'wlr_xdg_output.c',
-		'wlr_xdg_shell.c',
 	),
 	include_directories: wlr_inc,
 	dependencies: [pixman, xkbcommon, wayland_server, wlr_protos],
